serialize: generate hdf_serialize library
The hdf_serialize library is generated if hdf5 is found
on the system, and if it is not then no library is
generated. The library will be used to preserve daos
container properties and user attributes in daos filesystem
when data is moved from DAOS to a regular POSIX filesystem.
The library will also be used to store serialization and
deserialization functions for the future daos serialize
and daos deserialize utilities.

Signed-off-by: Danielle Sikich <danielle.sikich@intel.com>