DAOS-7767 container: create existence check by UUID and label
Before this change, a container create checked for existence of the
container by looking up the UUID in cs_conts KVS. However, with the
recently-added container open by_label API (and cs_uuids KVS), both
UUID and label need to be used to check container existence and check
the request for valid input.

This change adds a new cont_existence_check() function called early in
the cont_create() procedure to perform these checks.

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>